diff: cache diffing results
There are many cases when diffing a string where the actual start and expected start will
be the same as a previous actual start and expected start.  When that happens, the smallest
diff will be the same as the previous run.

It appears that the strings in #30 were re-calculating the smallest difference with the same
actual and expected starts over and over, resulting in effectively locking up the system.
Adding a map to cache results for a given run resolves the issue.