output cleanup for equal and differ
In matchers.Equal: use a full sentence for the comparison.  I believe that the original intent was for
the To type to include the word 'expected', but most matchers don't follow that pattern.  This updates
matchers.Equal to follow the same pattern as the other matchers, where the error message is the full
output for the test.

In differ.Differ.Diff: output the value of a pointer even if it's being compared to nil.  We were pretty
good about checking pointer values if they were being compared to other pointer values, but when comparing
to nil we would end up outputting the address of the non-nil pointer.  This solves that.