onpar: support generics in structured testing
This adds generics support to the structured framework of onpar, which provides
compile-time safety to parameter types in spec functions.  It seems like all
generic types are able to be inferred from arguments, so I'm actually pretty
happy with the state of generics here.

The one problem is that, since generics are not (yet) allowed to be defined as
part of method calls, we can't provide 'BeforeEach' as a method.  Instead, we're
providing it as a new constructor, which constructs and returns a new *OnPar.

For the most part, this seems to work fine ... but I can see how it might cause
some problems.  I'm hoping to solve some of those before merge.