onpar: use t.Cleanup to run the suite
With the advent of generics (and, therefor, 'BeforeEach' moving to a separate constructor),
the already-mildly-frustrating 'defer o.Run(t)' boilerplate became much worse.  Now, the
user would need to remember to add the boilerplate _after writing a BeforeEach setup function_.

I decided it was worth tracking down a better way, and finally found that 't.Cleanup' works
perfectly.  It allows 'onpar.New' to add the deferred 'o.Run(t)' itself, by setting it up
in the cleanup function.  When the test exits, it runs the suite as if the call to 'o.Run(t)'
had been deferred.