diff: add a timeout and string diff algorithm options
In weird edge cases, we keep running into situations where the diff hangs, trying to exhaust all possible diffs
and descending into near-infinite recursion.  To prevent this, we've added a timeout option and set the default
to time.Second.

To ensure that we return the best diff that we have available when the timer times out, we've also updated the
existing string diff algorithm to return results on a channel.  When the diff times out, we'll return the best
result that we've gotten from the channel.

There are many more things that could be improved, but this is already a hefty commit.