onpar: support go 1.20
As of go 1.20, calling t.Run from t.Cleanup panics. This breaks
onpar v2 and requires us to make a hard breaking change to
support.

The long story short is that 'defer o.Run()' is back, but onpar
will panic if o.Run() is not called before the test completes.